backend/services/cluster-manager-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
redis==5.0.1
python-jose[cryptography]==3.3.0
httpx==0.25.2
kubernetes==28.1.0
pyyaml==6.0.1
boto3==1.34.0
azure-identity==1.15.0
azure-mgmt-resource==23.0.1
google-cloud-resource-manager==1.10.4
google-auth==2.25.2
structlog==23.2.0
python-json-logger==2.0.7
psutil==5.9.6 